Fix man page path
The Makefile tries to detect /usr/share/man on the local system and, if not present, tries to install manual pages into /usr/man instead. Of course, this is not what we want, since `man` is not necessarily installed on the build machine, but the standard convention is still to install into /usr/share/man, therefore we'll just pretend that this path always exists.